🔵 moving helper funs to helpers.go
